No real changes, but
in the interest of increased and better communication, I have figured
out for myself (and possibly you if you didn't already know this) that
the middle test in test-basics is returning what appears to be an
environment instead of a result. I'm off to try and figure out what's
up.