More wild goose chasing print statements
So we do in fact have infinite loops, but only after the first time. So
something is getting changed the first time that shouldn't be. The first
time we have the error of incorrect argument count to iota. I made a
helper method for iota mostly so that I could do some more sanity checks
on it. I just can't figure out what's, oh wait. just had a thought. In
class today he mentioned that we would have to be careful if we had the
local environments extend the global environment, do you think that's
what he meant? Anyhow, I've been sitting at my desk for about 5 hours
now, and I think I may have a handle on where the problem is, but I'm
not sure, so feel free to email me if you want a crack at it.