Read these comments
The name of those test cases should have given you a hint: "Test if
set". The issue is that we're not checking to see if we already have the
item in the environment. The issue is more in apply-lambda. I've begun
working on a solution, but it is not complete yet and currently just
yields an infinite loop.

Work with Claude on this to see what is happening.